+++ packages/@webex/internal-plugin-metrics/test/unit/spec/call-diagnostic/call-diagnostic-metrics.util.ts
@@ -304,6 +304,7 @@ describe('internal-plugin-metrics', () => {
       ['client.login.end', {
         joinTimes: {
           otherAppApiReqResp: undefined,
+          exchangeCITokenJMT: undefined,
         }
       }],
       ['client.webexapp.launched', {